requirements.txt
accelerate==0.30.1
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
controlnet-aux==0.0.8
cycler==0.12.1
diffusers==0.22.1
einops==0.8.0
filelock==3.14.0
fonttools==4.53.0
fsspec==2024.5.0
huggingface-hub==0.23.3
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
ip-adapter @ git+https://github.com/tencent-ailab/IP-Adapter.git@11a904682126e5dfdd0c1ce86fdb435e9e01c1e0
Jinja2==3.1.4
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
psutil==5.9.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
safetensors==0.4.3
scikit-image==0.23.2
scipy==1.13.1
setuptools==69.5.1
six==1.16.0
sympy==1.12
tifffile==2024.5.22
timm==0.6.7
tokenizers==0.19.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.0
urllib3==2.2.1
wheel==0.43.0
zipp==3.19.0
